requirements.txt
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.2.0
idna==3.4
Pillow==10.0.0
pycryptodomex==3.18.0
requests==2.31.0
six==1.16.0
steam==1.4.4
urllib3==2.0.3
vdf==3.4
